created goods list, removed fetching goods array in goods list component
